refactor: 우승자 판단 로직 Referee 클래스로 분리
